trying to identify possinble 'needs mapping' ANs
part of possible need mapping part of the filter
function, the problem is that is not fully working
and we decided that is not that really important,
since those ANs will be just sent to the missing_dbxref
part of the pipeline, instead of missing_gene_dict. And
since most of them already have good gffs (with gene=, Note=),
that's is not a problem.