add the _from_class internal class method
this class method is a way to bypass `UserDict.__init__`,
when is called with an `inital`. The `__init__` will call self.update,
that in turn will run `self.data[key] = value` on each element of inital.

This is a bottleneck for large `inital` dicts. This bypass was made
internal, to be used only in situations where we are certain
that the `inital` data follow the structure expected by GeneDict.

These cases are `create_stripped()`, where the data are from self,
and the new function `rename_labels()`, where the data are also from
self.