new handling of GFF Paths
This new process to handle GFF path discovery is more flexible
in handling more cases of gff files index in the TSV.
Better for more complex pipelines that need different folder structures.
It allows for the user to maybe implement their on GFF path discovery,
like for instance an dictionary that will return a path using 'an' as
the key, as long as the class implemented work on a '__call__' execution.