great optimization on TEMP Description process
The main point here is the change to a producer/consumer
multiprocess, where for each batch of ANs, gff3 file reading,
parsing, ncbi query and merging are done on a producer worker
and inside a ProcessPool, returning the gff as df to the main thread
for further process.

Since the work done on query and adding if match to gene_dict cannot be
parallelized, due to its nature, this new process allows the main thread
to only do this work, not any other, making the process really fast.