`--large-temp` flag to stripped command
this will create a special type of stripped, to help
user parse large TEMP gdicts that have a ton of #dx entries.

since we are parsing a TEMP gdcit, we know that for each label
there'll be one #gd, either from from a gff_gene or NCBI. So first,
we save said #gd and all the ncbi_desc info from all #dx entries.
When we parse label data, we check if the first #gd is from a gff_gene,
if it is, we also keep all the unique ncbi descriptions that we found
in the label. If it's not (NCBI #gd), we just keep the #gd line.