better regex to clean up orfs/unknowns
before, we just removed features where`ID=` matched,
now we are also trying to remove any features that
are also connected to said orf, via the `Parent=`
connection.

since we are not filtering out based on coordinates,
some features may not be filtered out in gff3s that
are not well annotated.